address-list-view 모듈화 성공
